linux/x86_64/nlink_t: definition was wrong

+++ linux/x86_64/nlink_t.h
@@ -1,7 +1,7 @@
 #ifndef _c__linux__x86_64__nlink_t_h
 #define _c__linux__x86_64__nlink_t_h
 
-#include "c/uint32_t.h"
+#include "c/uint64_t.h"
 
 #ifdef __cplusplus
 inline namespace c {
@@ -9,7 +9,7 @@ inline namespace linux {
 inline namespace x86_64 {
 #endif
 
-typedef uint32_t nlink_t;
+typedef uint64_t nlink_t;
 
 #ifdef __cplusplus
 } // inline namespace x86_64
